# Code/Server/docs/swagger.yaml
basePath: /v1
definitions:
  db.Cleanliness:
    enum:
    - dirty
    - medium
    - clean
    type: string
    x-enum-varnames:
    - CleanlinessDirty
    - CleanlinessMedium
    - CleanlinessClean
...
  db.ReportType:
    enum:
    - start
    - middle
    - end
    type: string
    x-enum-varnames:
    - ReportTypeStart
    - ReportTypeMiddle
    - ReportTypeEnd
...
  db.State:
    enum:
    - broken
    - needsRepair
    - bad
    - medium
    - good
    - new
    type: string
    x-enum-varnames:
    - StateBroken
    - StateNeedsRepair
    - StateBad
    - StateMedium
    - StateGood
    - StateNew
...
  models.FurnitureStateResponse:
    properties:
      cleanliness:
        $ref: '#/definitions/db.Cleanliness'
      id:
        type: string
      name:
        type: string
      note:
        type: string
      pictures:
        items:
          type: string
        type: array
      quantity:
        type: integer
      state:
        $ref: '#/definitions/db.State'
    type: object
...
  models.InventoryReportResponse:
    properties:
      date:
        type: string
      id:
        type: string
      lease_id:
        type: string
      property_id:
        type: string
      rooms:
        items:
          $ref: '#/definitions/models.RoomStateResponse'
        type: array
      type:
        $ref: '#/definitions/db.ReportType'
    type: object
...
  models.RoomStateResponse:
    properties:
      cleanliness:
        $ref: '#/definitions/db.Cleanliness'
      furnitures:
        items:
          $ref: '#/definitions/models.FurnitureStateResponse'
        type: array
      id:
        type: string
      name:
        type: string
      note:
        type: string
      pictures:
        items:
          type: string
        type: array
      state:
        $ref: '#/definitions/db.State'
    type: object
  models.SummarizeRequest:
    properties:
      id:
        type: string
      pictures:
        items:
          type: string
        minItems: 1
        type: array
      type:
        enum:
        - room
        - furniture
        type: string
    required:
    - id
    - pictures
    - type
    type: object
  models.SummarizeResponse:
    properties:
      cleanliness:
        type: string
      note:
        type: string
      state:
        type: string
    type: object
...
  utils.Error:
    properties:
      code:
        $ref: '#/definitions/utils.ErrorCode'
      error:
        type: string
    type: object
  utils.ErrorCode:
    enum:
    - invalid-password
    - cannot-fetch-user
    - user-not-found
    - cannot-create-user
    - no-claims
    - cannot-decode-user
    - missing-fields
    - missing-file
    - cannot-hash-password
    - email-already-exists
    - test-error
    - too-many-requests
    - invite-not-found
    - user-must-have-same-email-as-invite
    - invite-already-exists-for-email-or-property
    - lease-already-exists-for-tenant-and-property
    - property-not-found
    - property-is-not-yours
    - not-an-owner
    - not-a-tenant
    - user-already-exists-as-owner
    - property-already-exists
    - property-not-available
    - tenant-already-has-lease
    - no-active-lease
    - lease-not-found
    - cannot-end-non-current-lease
    - inventory-report-must-be-linked-to-current-lease
    - no-pending-lease
    - document-not-found
    - damage-not-found
    - damage-already-exists
    - cannot-update-fixed-damage
    - damage-already-fixed
    - failed-upload-file
    - failed-delete-file
    - bad-base64-string
    - property-picture-not-found
    - user-profile-picture-not-found
    - room-not-found
    - room-already-exists
    - room-not-in-this-property
    - furniture-not-found
    - furniture-already-exists
    - furniture-not-in-this-property
    - inventory-report-already-exists
    - cannot-modify-inventory-report
    - inventory-report-not-found
    - room-state-already-exists
    - furniture-state-already-exists
    - invalid-object-type
    - error-request-chatgpt-api
    - failed-send-email
    type: string
    x-enum-varnames:
    - InvalidPassword
    - CannotFetchUser
    - UserNotFound
    - CannotCreateUser
    - NoClaims
    - CannotDecodeUser
    - MissingFields
    - MissingFile
    - CannotHashPassword
    - EmailAlreadyExists
    - TestError
    - TooManyRequests
    - InviteNotFound
    - UserSameEmailAsInvite
    - InviteAlreadyExists
    - LeaseAlreadyExist
    - PropertyNotFound
    - PropertyNotYours
    - NotAnOwner
    - NotATenant
    - UserAlreadyExistsAsOwner
    - PropertyAlreadyExists
    - PropertyNotAvailable
    - TenantAlreadyHasLease
    - NoActiveLease
    - LeaseNotFound
    - CannotEndNonCurrentLease
    - InvReportMustBeCurrentLease
    - NoLeaseInvite
    - DocumentNotFound
    - DamageNotFound
    - DamageAlreadyExists
    - CannotUpdateFixedDamage
    - DamageAlreadyFixed
    - FailedUploadFile
    - FailedDeleteFile
    - BadBase64String
    - PropertyPictureNotFound
    - UserProfilePictureNotFound
    - RoomNotFound
    - RoomAlreadyExists
    - RoomNotInThisProperty
    - FurnitureNotFound
    - FurnitureAlreadyExists
    - FurnitureNotInThisProperty
    - InventoryReportAlreadyExists
    - CannotModifyInventoryReport
    - InventoryReportNotFound
    - RoomStateAlreadyExists
    - FurnitureStateAlreadyExists
    - InvalidObjectType
    - ErrorRequestChatGPTAPI
    - FailedSendEmail
host: localhost:3001
info:
  contact:
    email: martin.d-herouville@epitech.eu
    name: Mazettt
  description: This is the API used by the Immotep application.
  termsOfService: http://swagger.io/terms/
  title: Immotep API
  version: "1.0"
paths:
...
  /owner/properties/{property_id}/leases/{lease_id}/inventory-reports/compare/{old_report_id}/:
    post:
      consumes:
      - application/json
      description: Generate comparison from photo for last inventory report
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Previous report ID to compare with
        in: path
        name: old_report_id
        required: true
        type: string
      - description: Compare data
        in: body
        name: aiData
        required: true
        schema:
          $ref: '#/definitions/models.SummarizeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created inventory report data
          schema:
            $ref: '#/definitions/models.InventoryReportResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property or old report not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Generate comparison from photo
      tags:
      - ai
  /owner/properties/{property_id}/leases/{lease_id}/inventory-reports/summarize/:
    post:
      consumes:
      - application/json
      description: Generate summary from photo for first inventory report
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Summarize data
        in: body
        name: aiData
        required: true
        schema:
          $ref: '#/definitions/models.SummarizeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Summarize data
          schema:
            $ref: '#/definitions/models.SummarizeResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "403":
          description: Property not yours
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Generate summary from photo
      tags:
      - ai
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"